+++ lib/iex/errors/error.rb
@@ -1,0 +1,5 @@
+module IEX
+  module Errors
+    class Error < StandardError; end
+  end
+end