+++ lib/iex/errors/search_not_found_error.rb
@@ -1,0 +1,16 @@
+require_relative 'error'
+
+module IEX
+  module Errors
+    class SearchNotFoundError < Error
+      attr_reader :response
+      attr_reader :fragment
+
+      def initialize(fragment, response)
+        @response = response
+        @fragment = fragment
+        super %(Fragment "#{fragment}" Not Found)
+      end
+    end
+  end
+end